+++ winrm/transport.py
@@ -1,18 +1,14 @@
 from __future__ import unicode_literals
-from contextlib import contextmanager
-import re
 import sys
 import os
-import weakref
+import inspect
 
 is_py2 = sys.version[0] == '2'
 
 if is_py2:
-    from urlparse import urlsplit, urlunsplit
     # use six for this instead?
     unicode_type = type(u'')
 else:
-    from urllib.parse import urlsplit, urlunsplit
     # use six for this instead?
     unicode_type = type(u'')
 
@@ -20,12 +16,11 @@ import requests
 import requests.auth
 import warnings
 from distutils.util import strtobool
-from requests.hooks import default_hooks
-from requests.adapters import HTTPAdapter
 
 HAVE_KERBEROS = False
 try:
     from requests_kerberos import HTTPKerberosAuth, REQUIRED, OPTIONAL, DISABLED
+
     HAVE_KERBEROS = True
 except ImportError:
     pass
@@ -33,6 +28,7 @@ except ImportError:
 HAVE_NTLM = False
 try:
     from requests_ntlm import HttpNtlmAuth
+
     HAVE_NTLM = True
 except ImportError as ie:
     pass
@@ -40,26 +36,33 @@ except ImportError as ie:
 HAVE_CREDSSP = False
 try:
     from requests_credssp import HttpCredSSPAuth
+
     HAVE_CREDSSP = True
 except ImportError as ie:
     pass
 
-from winrm.exceptions import BasicAuthDisabledError, InvalidCredentialsError, \
-    WinRMError, WinRMTransportError, WinRMOperationTimeoutError
+from winrm.exceptions import InvalidCredentialsError, WinRMError, \
+    WinRMTransportError
+from winrm.encryption import Encryption
 
 __all__ = ['Transport']
 
-import ssl
+
+class UnsupportedAuthArgument(Warning):
+    pass
+
 
 class Transport(object):
-    
     def __init__(
             self, endpoint, username=None, password=None, realm=None,
             service=None, keytab=None, ca_trust_path=None, cert_pem=None,
             cert_key_pem=None, read_timeout_sec=None, server_cert_validation='validate',
             kerberos_delegation=False,
             kerberos_hostname_override=None,
-            auth_method='auto'):
+            auth_method='auto',
+            message_encryption='auto',
+            credssp_disable_tlsv1_2=False,
+            send_cbt=True):
         self.endpoint = endpoint
         self.username = username
         self.password = password
@@ -72,6 +75,9 @@ class Transport(object):
         self.read_timeout_sec = read_timeout_sec
         self.server_cert_validation = server_cert_validation
         self.kerberos_hostname_override = kerberos_hostname_override
+        self.message_encryption = message_encryption
+        self.credssp_disable_tlsv1_2 = credssp_disable_tlsv1_2
+        self.send_cbt = send_cbt
 
         if self.server_cert_validation not in [None, 'validate', 'ignore']:
             raise WinRMError('invalid server_cert_validation mode: %s' % self.server_cert_validation)
@@ -92,12 +98,14 @@ class Transport(object):
         try:
             from requests.packages.urllib3.exceptions import InsecurePlatformWarning
             warnings.simplefilter('ignore', category=InsecurePlatformWarning)
-        except: pass # oh well, we tried...
+        except:
+            pass  # oh well, we tried...
 
         try:
             from requests.packages.urllib3.exceptions import SNIMissingWarning
             warnings.simplefilter('ignore', category=SNIMissingWarning)
-        except: pass # oh well, we tried...
+        except:
+            pass  # oh well, we tried...
 
         # if we're explicitly ignoring validation, try to suppress InsecureRequestWarning, since the user opted-in
         if self.server_cert_validation == 'ignore':
@@ -130,10 +138,18 @@ class Transport(object):
 
         self.session = None
 
+        # Used for encrypting messages
+        self.encryption = None  # The Pywinrm Encryption class used to encrypt/decrypt messages
+        if self.message_encryption not in ['auto', 'always', 'never']:
+            raise WinRMError(
+                "invalid message_encryption arg: %s. Should be 'auto', 'always', or 'never'" % self.message_encryption)
+
     def build_session(self):
         session = requests.Session()
 
         session.verify = self.server_cert_validation == 'validate'
+        if session.verify and self.ca_trust_path:
+                session.verify = self.ca_trust_path
 
         # configure proxies from HTTP/HTTPS_PROXY envvars
         session.trust_env = True
@@ -143,15 +159,27 @@ class Transport(object):
         # we're only applying proxies from env, other settings are ignored
         session.proxies = settings['proxies']
 
+        encryption_available = False
         if self.auth_method == 'kerberos':
             if not HAVE_KERBEROS:
                 raise WinRMError("requested auth method is kerberos, but requests_kerberos is not installed")
-            # TODO: do argspec sniffing on extensions to ensure we're not setting bogus kwargs on older versions
-            session.auth = HTTPKerberosAuth(mutual_authentication=REQUIRED, delegate=self.kerberos_delegation,
-                                            force_preemptive=True, principal=self.username,
-                                            hostname_override=self.kerberos_hostname_override,
-                                            sanitize_mutual_error_response=False)
-        elif self.auth_method in ['certificate','ssl']:
+
+            man_args = dict(
+                mutual_authentication=REQUIRED,
+            )
+            opt_args = dict(
+                delegate=self.kerberos_delegation,
+                force_preemptive=True,
+                principal=self.username,
+                hostname_override=self.kerberos_hostname_override,
+                sanitize_mutual_error_response=False,
+                service=self.service,
+                send_cbt=self.send_cbt
+            )
+            kerb_args = self._get_args(man_args, opt_args, HTTPKerberosAuth.__init__)
+            session.auth = HTTPKerberosAuth(**kerb_args)
+            encryption_available = hasattr(session.auth, 'winrm_encryption_available') and session.auth.winrm_encryption_available
+        elif self.auth_method in ['certificate', 'ssl']:
             if self.auth_method == 'ssl' and not self.cert_pem and not self.cert_key_pem:
                 # 'ssl' was overloaded for HTTPS with optional certificate auth,
                 # fall back to basic auth if no cert specified
@@ -163,53 +191,107 @@ class Transport(object):
         elif self.auth_method == 'ntlm':
             if not HAVE_NTLM:
                 raise WinRMError("requested auth method is ntlm, but requests_ntlm is not installed")
-            session.auth = HttpNtlmAuth(username=self.username, password=self.password)
+            man_args = dict(
+                username=self.username,
+                password=self.password
+            )
+            opt_args = dict(
+                send_cbt=self.send_cbt
+            )
+            ntlm_args = self._get_args(man_args, opt_args, HttpNtlmAuth.__init__)
+            session.auth = HttpNtlmAuth(**ntlm_args)
+            # check if requests_ntlm has the session_security attribute available for encryption
+            encryption_available = hasattr(session.auth, 'session_security')
         # TODO: ssl is not exactly right here- should really be client_cert
-        elif self.auth_method in ['basic','plaintext']:
+        elif self.auth_method in ['basic', 'plaintext']:
             session.auth = requests.auth.HTTPBasicAuth(username=self.username, password=self.password)
         elif self.auth_method == 'credssp':
             if not HAVE_CREDSSP:
                 raise WinRMError("requests auth method is credssp, but requests-credssp is not installed")
-            session.auth = HttpCredSSPAuth(username=self.username, password=self.password)
-
+            session.auth = HttpCredSSPAuth(username=self.username, password=self.password,
+                                               disable_tlsv1_2=self.credssp_disable_tlsv1_2)
+            encryption_available = hasattr(session.auth, 'wrap') and hasattr(session.auth, 'unwrap')
         else:
             raise WinRMError("unsupported auth method: %s" % self.auth_method)
 
         session.headers.update(self.default_headers)
+        self.session = session
 
-        return session
+        # Will check the current config and see if we need to setup message encryption
+        if self.message_encryption == 'always' and not encryption_available:
+            raise WinRMError(
+                "message encryption is set to 'always' but the selected auth method %s does not support it" % self.auth_method)
+        elif encryption_available:
+            if self.message_encryption == 'always':
+                self.setup_encryption()
+            elif self.message_encryption == 'auto' and not self.endpoint.lower().startswith('https'):
+                self.setup_encryption()
+
+    def setup_encryption(self):
+        # Security context doesn't exist, sending blank message to initialise context
+        request = requests.Request('POST', self.endpoint, data=None)
+        prepared_request = self.session.prepare_request(request)
+        self._send_message_request(prepared_request, '')
+        self.encryption = Encryption(self.session, self.auth_method)
 
     def send_message(self, message):
-        # TODO support kerberos/ntlm session with message encryption
-
         if not self.session:
-            self.session = self.build_session()
+            self.build_session()
 
         # urllib3 fails on SSL retries with unicode buffers- must send it a byte string
         # see https://github.com/shazow/urllib3/issues/717
         if isinstance(message, unicode_type):
             message = message.encode('utf-8')
 
-        request = requests.Request('POST', self.endpoint, data=message)
-        prepared_request = self.session.prepare_request(request)
+        if self.encryption:
+            prepared_request = self.encryption.prepare_encrypted_request(self.session, self.endpoint, message)
+        else:
+            request = requests.Request('POST', self.endpoint, data=message)
+            prepared_request = self.session.prepare_request(request)
 
+        response = self._send_message_request(prepared_request, message)
+        return self._get_message_response_text(response)
+
+    def _send_message_request(self, prepared_request, message):
         try:
             response = self.session.send(prepared_request, timeout=self.read_timeout_sec)
-            response_text = response.text
             response.raise_for_status()
-            return response_text
+            return response
         except requests.HTTPError as ex:
             if ex.response.status_code == 401:
                 raise InvalidCredentialsError("the specified credentials were rejected by the server")
             if ex.response.content:
-                response_text = ex.response.content
+                response_text = self._get_message_response_text(ex.response)
             else:
                 response_text = ''
-            # Per http://msdn.microsoft.com/en-us/library/cc251676.aspx rule 3,
-            # should handle this 500 error and retry receiving command output.
-            if b'http://schemas.microsoft.com/wbem/wsman/1/windows/shell/Receive' in message and b'Code="2150858793"' in response_text:
-                raise WinRMOperationTimeoutError()
 
-            error_message = 'Bad HTTP response returned from server. Code {0}'.format(ex.response.status_code)
 
-            raise WinRMTransportError('http', error_message)
+            raise WinRMTransportError('http', ex.response.status_code, response_text)
+
+    def _get_message_response_text(self, response):
+        if self.encryption:
+            response_text = self.encryption.parse_encrypted_response(response)
+        else:
+            response_text = response.content
+        return response_text
+
+    def _get_args(self, mandatory_args, optional_args, function):
+        argspec = set(inspect.getargspec(function).args)
+        function_args = dict()
+        for name, value in mandatory_args.items():
+            if name in argspec:
+                function_args[name] = value
+            else:
+                raise Exception("Function %s does not contain mandatory arg "
+                                "%s, check installed version with pip list"
+                                % (str(function), name))
+
+        for name, value in optional_args.items():
+            if name in argspec:
+                function_args[name] = value
+            else:
+                warnings.warn("Function %s does not contain optional arg %s, "
+                              "check installed version with pip list"
+                              % (str(function), name))
+
+        return function_args