+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Developing
+- Added support for message encryption over HTTP when using NTLM and CredSSP
+- Added parameter to disable TLSv1.2 when using CredSSP for Server 2008 support
+
 ### Version 0.2.2
 - Added support for CredSSP authenication (via requests-credssp)
 - Improved README, see 'Valid transport options' section